backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
stripe==6.7.0
PyJWT==2.8.0
bcrypt==4.0.1
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
requests==2.31.0
openai==1.3.5
replicate==0.15.4
django-storages==1.14.2
boto3==1.29.7
sendgrid==6.10.0
django-extensions==3.2.3
